Re-factored user mongo tests to pass; the only broken tests was the findAll where the id was missing in two of the three records being created.
